feat: add a confirm pop-up on proposition deletion

+++ src/main/front/src/pages/proposal/edit/ProposalPage.js
@@ -129,6 +129,10 @@ class ProposalPage extends Component {
     };
 
     onDelete = (id) => {
+
+        const wantToDelete = window.confirm("Do you really want to delete this proposition ?")
+        if(!wantToDelete) return;
+
         axios.delete(`/api/proposal/${this.state.id}/choices/${id}`)
             .then(response => {
                 this.setState({choices: response.data.choices})
